refactor: clearer separation for fallback components
